Connections Lines and Government Influence
Readded government and government influence and growth. Created a
territory effect and connections options for the galaxy view. The info
does seem to save but need to fix the ui issue on load.

+++ Trader/Assets/_Game/Scripts/Models/SolarModel.cs
@@ -21,7 +21,7 @@ public class SolarModel : Model {
     public ModelRefs<ShipModel> ships = new ModelRefs<ShipModel>();
     public ModelRefs<SolarModel> nearStars = new ModelRefs<SolarModel>();
     public ModelRef<GovernmentModel> government = new ModelRef<GovernmentModel>();
-    public float governmentInfluence;
+    public double governmentInfluence;
     public bool isCapital;
     public int index;
     internal float localScale;
